# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Air France PTFS API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type FlightStatus = typeof FlightStatus[keyof typeof FlightStatus];


export const FlightStatus = {
  scheduled: 'scheduled',
  boarding: 'boarding',
  departed: 'departed',
  arrived: 'arrived',
  cancelled: 'cancelled',
} as const;

export interface Flight {
  id: number;
  flightNumber: string;
  origin: string;
  destination: string;
  departureTime: string;
  arrivalTime: string;
  aircraft: string;
  status: FlightStatus;
  seats: number;
  bookedSeats: number;
  price: number;
  /** @nullable */
  imageUrl?: string | null;
}

export interface FlightInput {
  flightNumber: string;
  origin: string;
  destination: string;
  departureTime: string;
  arrivalTime: string;
  aircraft: string;
  seats: number;
  price: number;
  imageUrl?: string;
}

export type FlightUpdateStatus = typeof FlightUpdateStatus[keyof typeof FlightUpdateStatus];


export const FlightUpdateStatus = {
  scheduled: 'scheduled',
  boarding: 'boarding',
  departed: 'departed',
  arrived: 'arrived',
  cancelled: 'cancelled',
} as const;

export interface FlightUpdate {
  flightNumber?: string;
  origin?: string;
  destination?: string;
  departureTime?: string;
  arrivalTime?: string;
  aircraft?: string;
  status?: FlightUpdateStatus;
  seats?: number;
  price?: number;
  imageUrl?: string;
}

export type BookingSeatClass = typeof BookingSeatClass[keyof typeof BookingSeatClass];


export const BookingSeatClass = {
  economy: 'economy',
  business: 'business',
  first: 'first',
} as const;

export type BookingStatus = typeof BookingStatus[keyof typeof BookingStatus];


export const BookingStatus = {
  confirmed: 'confirmed',
  cancelled: 'cancelled',
  pending: 'pending',
} as const;

export interface Booking {
  id: number;
  flightId: number;
  flight?: Flight;
  passengerName: string;
  seatClass: BookingSeatClass;
  status: BookingStatus;
  createdAt: string;
}

export type BookingInputSeatClass = typeof BookingInputSeatClass[keyof typeof BookingInputSeatClass];


export const BookingInputSeatClass = {
  economy: 'economy',
  business: 'business',
  first: 'first',
} as const;

export interface BookingInput {
  flightId: number;
  passengerName: string;
  seatClass: BookingInputSeatClass;
}

export type ProfileRole = typeof ProfileRole[keyof typeof ProfileRole];


export const ProfileRole = {
  passenger: 'passenger',
  crew: 'crew',
  admin: 'admin',
} as const;

export interface Profile {
  id: number;
  username: string;
  role: ProfileRole;
  joinedAt: string;
  totalFlights: number;
  memberSince: string;
  /** @nullable */
  avatarUrl?: string | null;
  /** @nullable */
  bio?: string | null;
}

export interface ProfileUpdate {
  username?: string;
  bio?: string;
  avatarUrl?: string;
}

export interface AdminStats {
  totalFlights: number;
  totalBookings: number;
  activePassengers: number;
  scheduledFlights: number;
  cancelledFlights: number;
  totalRevenue: number;
}