# lib/api-zod/src/generated/types/adminStats.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Air France PTFS API
 * OpenAPI spec version: 0.1.0
 */

export interface AdminStats {
  totalFlights: number;
  totalBookings: number;
  activePassengers: number;
  scheduledFlights: number;
  cancelledFlights: number;
  totalRevenue: number;
}