# lib/api-zod/src/generated/types/booking.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Air France PTFS API
 * OpenAPI spec version: 0.1.0
 */
import type { BookingSeatClass } from './bookingSeatClass';
import type { BookingStatus } from './bookingStatus';
import type { Flight } from './flight';

export interface Booking {
  id: number;
  flightId: number;
  flight?: Flight;
  passengerName: string;
  seatClass: BookingSeatClass;
  status: BookingStatus;
  createdAt: string;
}