# lib/api-zod/src/generated/types/bookingInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Air France PTFS API
 * OpenAPI spec version: 0.1.0
 */
import type { BookingInputSeatClass } from './bookingInputSeatClass';

export interface BookingInput {
  flightId: number;
  passengerName: string;
  seatClass: BookingInputSeatClass;
}