# lib/api-zod/src/generated/types/flight.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Air France PTFS API
 * OpenAPI spec version: 0.1.0
 */
import type { FlightStatus } from './flightStatus';

export interface Flight {
  id: number;
  flightNumber: string;
  origin: string;
  destination: string;
  departureTime: string;
  arrivalTime: string;
  aircraft: string;
  status: FlightStatus;
  seats: number;
  bookedSeats: number;
  price: number;
  /** @nullable */
  imageUrl?: string | null;
}