# lib/api-zod/src/generated/types/flightInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Air France PTFS API
 * OpenAPI spec version: 0.1.0
 */

export interface FlightInput {
  flightNumber: string;
  origin: string;
  destination: string;
  departureTime: string;
  arrivalTime: string;
  aircraft: string;
  seats: number;
  price: number;
  imageUrl?: string;
}