# lib/api-zod/src/generated/types/flightUpdate.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Air France PTFS API
 * OpenAPI spec version: 0.1.0
 */
import type { FlightUpdateStatus } from './flightUpdateStatus';

export interface FlightUpdate {
  flightNumber?: string;
  origin?: string;
  destination?: string;
  departureTime?: string;
  arrivalTime?: string;
  aircraft?: string;
  status?: FlightUpdateStatus;
  seats?: number;
  price?: number;
  imageUrl?: string;
}