# lib/api-zod/src/generated/types/profile.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Air France PTFS API
 * OpenAPI spec version: 0.1.0
 */
import type { ProfileRole } from './profileRole';

export interface Profile {
  id: number;
  username: string;
  role: ProfileRole;
  joinedAt: string;
  totalFlights: number;
  memberSince: string;
  /** @nullable */
  avatarUrl?: string | null;
  /** @nullable */
  bio?: string | null;
}