# lib/api-zod/src/generated/types/profileUpdate.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Air France PTFS API
 * OpenAPI spec version: 0.1.0
 */

export interface ProfileUpdate {
  username?: string;
  bio?: string;
  avatarUrl?: string;
}